Turned amount and exchange rates from floats into BigDecimals
I have not tried upgrading a plugin in a live system, and this will
likely have implications for the database, so bear that in mind
before you upgrade.

+++ grails-app/domain/cr/co/arquetipos/currencies/ExchangeRate.groovy
@@ -2,7 +2,7 @@ package cr.co.arquetipos.currencies
 class ExchangeRate implements Serializable { 
     Currency baseCurrency
     Currency toCurrency
-    float rate
+    BigDecimal rate
     Date date = new Date()
     
     static constraints = {
